+++ src/cards/CMakeLists.txt
@@ -6,6 +6,7 @@ add_library(cards_lib STATIC
     CardSet.cpp
     Deal.cpp
     FourHands.cpp
+    utils.cpp
 )
 
 target_link_libraries(cards_lib
