+++ src/cards/utils.cpp
@@ -1,4 +1,4 @@
-#include "gstate/utils.hpp"
+#include "cards/utils.hpp"
 #include "math/random.hpp"
 #include "prim/range.hpp"
