+++ src/gstate/GState.cpp
@@ -1,5 +1,5 @@
 #include "gstate/GState.hpp"
-#include "gstate/utils.hpp"
+#include "cards/utils.hpp"
 #include "prim/range.hpp"
 #include "math/Int128.hpp"
