+++ src/gstate/tests/GameOutcome.cpp
@@ -2,7 +2,7 @@
 
 #include "gstate/GameBehavior.hpp"
 #include "gstate/GState.hpp"
-#include "gstate/utils.hpp"
+#include "cards/utils.hpp"
 #include "prim/range.hpp"
 #include "stats/RunningStats.hpp"
